# api/openapi.yaml
openapi: 3.1.0
info:
  title: Cooked API
  version: 1.0.0-m1
  description: Contract for the Cooked v1 API. Product behavior is defined in product-frd.md.
servers:
  - url: /api/v1
...
paths:
...
  /auth/verify-email:
    get:
      tags: [Authentication]
      operationId: verifyEmail
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Email verified.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /auth/logout-all:
    post:
      tags: [Authentication]
      operationId: logoutAll
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: All token-family sessions revoked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/forgot-password:
    post:
      tags: [Authentication]
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "200":
          description: Generic accepted response, regardless of account existence.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /auth/reset-password:
    post:
      tags: [Authentication]
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password reset and outstanding reset credentials invalidated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication failed or is required.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    TooManyRequests:
      description: Rate limit exceeded.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    Meta:
      type: object
      required: [timestamp]
      properties:
        request_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        next_cursor:
          type: string
    SuccessResponse:
      type: object
      required: [status, meta, data]
      properties:
        status:
          const: success
        meta:
          $ref: "#/components/schemas/Meta"
        data: {}
    ErrorResponse:
      type: object
      required: [status, meta, error]
      properties:
        status:
          const: error
        meta:
          $ref: "#/components/schemas/Meta"
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details: {}
    User:
      type: object
      required: [id, email, name, user_name, is_verified, roles]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        user_name:
          type: string
        is_verified:
          type: boolean
        picture:
          type: string
          format: uri
        roles:
          type: array
          uniqueItems: true
          items:
            type: string
            enum: [user, moderator, admin]
        bio: {type: [string, "null"], maxLength: 1024}
        avatar_media_id: {type: [string, "null"], format: uuid}
        timezone: {type: string, description: IANA timezone name}
        dietary_preferences:
          type: array
          items: {$ref: "#/components/schemas/DietaryTag"}
        xp_total: {type: integer, minimum: 0}
        current_streak: {type: integer, minimum: 0}
        longest_streak: {type: integer, minimum: 0}
    DietaryTag:
      type: object
      required: [id, name, slug]
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        slug: {type: string, enum: [vegetarian, vegan, halal, pescatarian]}
...
    RefreshTokenRequest:
      type: object
      additionalProperties: false
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          minLength: 32
    ResetPasswordRequest:
      type: object
      additionalProperties: false
      required: [otp, email, password]
      properties:
        otp:
          type: string
          pattern: "^[0-9]{6}$"
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
...
    UserResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          properties:
            data:
              type: object
              required: [user, message]
              properties:
                user:
                  $ref: "#/components/schemas/User"
                message:
                  type: string
...